Prevent users from deactivating Ranger rule
As of Spark 2.4, there's a new configuration
(`spark.sql.optimizer.excludedRules`) which allows to excluide some
rules from the optimizer.

This project injects some optimizer rules in order to authorize with
Ranger rules. This caused a security issue where some user could
execute:
`SET spark.sql.optimizer.excludedRules=org.apache.spark.sql.catalyst.optimizer.RangerSparkAuthorizerExtension`

This disables the authorizing rule and allows to bypass all Ranger
rules.

To prevent that, this rule will check if someone tries to disable any
rule which contains the `Ranger` string and will throw an exception. The
unique way to workarround this is that this configuration is set at
Spark level. In this case, we consider it a bad configuration of the
cluster operator.